// src/types/enums.ts
export enum homeReducerEnums {
    GET_SUCCESS_GREETINGS = 'GET_SUCCESS_GREETINGS',
    GET_SUCCESS_SOCIAL_MEDIA = 'GET_SUCCESS_SOCIAL_MEDIA',
    GET_LOADING = "GET_LOADING",
    GET_ERROR = "GET_ERROR",
    GET_ALL_HOME_DATA = "GET_ALL_HOME_DATA"
}

export enum SideBarPayloadPageNameEnum {
    HOME = 'HOME',
    EDUCATION = 'EDUCATION',
    EXPERIENCE = 'EXPERIENCE',
    PROJECTS = 'PROJECTS',
    OPEN_SOURCE = 'OPEN_SOURCE',
    CONTACT_ME = 'CONTACT_ME',
}

export enum SideBarReducerPageNameEnum {
    SIDEBAR_LOADING = 'SIDEBAR_LOADING',
    SIDEBAR_ERROR = 'SIDEBAR_ERROR',
    SELECTED_PAGE = 'SELECTED_PAGE',
}